extract metadata validation
