+++ extensions/sendbird/v1/actions/createUser/config/fields.ts
@@ -1,7 +1,6 @@
 import { z, type ZodTypeAny } from 'zod'
-import { isEmpty, isNil } from 'lodash'
 import { type Field, FieldType } from '@awell-health/extensions-core'
-import { type Metadata } from '../../../types'
+import { MetadataValidationSchema } from '../../../validation'
 
 export const fields = {
   userId: {
@@ -41,46 +40,5 @@ export const FieldsValidationSchema = z.object({
   userId: z.string().max(80).nonempty(),
   nickname: z.string().max(80).nonempty(),
   issueAccessToken: z.boolean().optional(),
-  metadata: z.optional(z.string()).transform((str, ctx): Metadata => {
-    if (isNil(str) || isEmpty(str)) return {}
-
-    try {
-      const parsedJson = JSON.parse(str)
-
-      if (isEmpty(parsedJson)) {
-        return {}
-      }
-
-      if (typeof parsedJson !== 'object' || Array.isArray(parsedJson)) {
-        ctx.addIssue({
-          code: 'custom',
-          message: 'metadata should be an object',
-        })
-        return z.NEVER
-      }
-
-      const values = Object.values(parsedJson)
-
-      if (values.length > 5) {
-        ctx.addIssue({
-          code: 'custom',
-          message: 'metadata should have maximum of five key-value items',
-        })
-        return z.NEVER
-      }
-
-      if (values.some((val) => typeof val !== 'string')) {
-        ctx.addIssue({
-          code: 'custom',
-          message: 'The value of each metadata key must be a string',
-        })
-        return z.NEVER
-      }
-
-      return parsedJson
-    } catch (e) {
-      ctx.addIssue({ code: 'custom', message: 'Invalid JSON' })
-      return z.NEVER
-    }
-  }),
+  metadata: z.optional(MetadataValidationSchema),
 } satisfies Record<keyof typeof fields, ZodTypeAny>)
